feat(src): :sparkles: now BaseNature can auto call its major layer's mehtods
